Initialize reanimated context in Storybook
